Order Service: Get order details, refactor error model

+++ order-service/src/main/kotlin/com/codergm/orderser/domain/dto/OrderDto.kt
@@ -5,7 +5,7 @@ import java.time.Instant
 
 data class OrderDto(val id: Long?,
                     val productId: Long,
-                    val quantity: Long,
+                    val quantity: Int,
                     val orderDate: Instant,
                     val orderStatus: OrderStatus,
                     val amount: Double)
